scripts/requirements.txt
requests
BeautifulSoup4
paho-mqtt
